refactor: splited nulltype.py and renamed NullSafeDict to qdict
